requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.28.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
groq>=0.4.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
Pillow>=10.0.0
altair>=5.0.0
statsmodels>=0.14.0
scipy>=1.11.0
